requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-dotenv==0.21.0
mysqlclient==2.2.4
